A few changes in FFT tests, see below
1. Skip the allocate test due to a cuFFT bug (#3777)
2. Fix @multi_gpu
3. Split callback tests
